(Backup push) Spot lights working (need to tweak light model, for better performance)

+++ Paradime 3D/Clock.cpp
@@ -84,7 +84,7 @@ float Clock::GetMSfloat()
 
 int Clock::GetFPS()
 {
-	return (1000 / lastFrameMS);
+	return (int)(1000 / lastFrameMS);
 }
 
 string Clock::GetFPSstring()
